# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReplyX API",
    "description": "\nReplyX - Comprehensive AI Assistant Platform API\n\n## Features\n- 🤖 AI Assistant Management\n- 💬 Real-time Dialog System  \n- 📄 Document Processing & Knowledge Base\n- 🤖 Telegram Bot Integration\n- 💰 Balance & Billing System\n- 👥 User Management & Authentication\n- 📊 Analytics & Monitoring\n- 🛡️ Security & Rate Limiting\n\n## Authentication\nAll endpoints (except public ones) require JWT authentication:\n- Include `Authorization: Bearer <token>` header\n- Include `X-CSRF-Token` header for state-changing operations\n\n## Base URLs\n- Production: `https://api.replyx.ru`\n- Development: `http://localhost:8000`\n        ",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.replyx.ru",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/dialogs/{dialog_id}/handoff/request": {
      "post": {
        "tags": [
          "handoff"
        ],
        "summary": "Request Handoff",
        "description": "Request operator handoff with idempotency protection.\n\n- **dialog_id**: ID of the dialog to request handoff for\n- **body**: Request details including reason and request_id for idempotency\n\nReturns current handoff status. If request_id was already processed,\nreturns the same response (idempotency).\n\n**Rate limits**: Max 3 requests per minute per dialog.",
        "operationId": "request_handoff_api_dialogs__dialog_id__handoff_request_post",
        "parameters": [
          {
            "name": "dialog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dialog Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffRequestIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dialogs/{dialog_id}/handoff/takeover": {
      "post": {
        "tags": [
          "handoff"
        ],
        "summary": "Takeover Handoff",
        "description": "Operator takes over dialog with concurrency protection.\n\n- **dialog_id**: ID of the dialog to take over\n- Manager ID is extracted from JWT token\n\nOnly works if dialog is in 'requested' state and operator has capacity.\nUses SELECT FOR UPDATE to prevent race conditions.",
        "operationId": "takeover_handoff_api_dialogs__dialog_id__handoff_takeover_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dialog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dialog Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffTakeoverIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dialogs/{dialog_id}/handoff/release": {
      "post": {
        "tags": [
          "handoff"
        ],
        "summary": "Release Handoff",
        "description": "Release dialog back to AI (only by assigned manager).\n\n- **dialog_id**: ID of the dialog to release\n- Only the assigned manager can release the dialog\n- Transitions dialog from 'active' to 'released' state\n- Re-enables AI responses and decreases operator load",
        "operationId": "release_handoff_api_dialogs__dialog_id__handoff_release_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dialog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dialog Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffReleaseIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dialogs/{dialog_id}/handoff/cancel": {
      "post": {
        "tags": [
          "handoff"
        ],
        "summary": "Cancel Handoff",
        "description": "Cancel handoff request (before it becomes active).\n\n- **dialog_id**: ID of the dialog to cancel handoff for\n- Can only cancel requests in 'requested' state\n- Transitions dialog to 'cancelled' state",
        "operationId": "cancel_handoff_api_dialogs__dialog_id__handoff_cancel_post",
        "parameters": [
          {
            "name": "dialog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dialog Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffCancelIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dialogs/{dialog_id}/handoff/status": {
      "get": {
        "tags": [
          "handoff"
        ],
        "summary": "Get Handoff Status",
        "description": "Get current handoff status for UI synchronization.\n\n- **dialog_id**: ID of the dialog to get status for\n\nReturns complete handoff status including assigned manager,\nqueue position, estimated wait time, and timestamps.\nUsed by frontend to sync UI state after WebSocket reconnection.",
        "operationId": "get_handoff_status_api_dialogs__dialog_id__handoff_status_get",
        "parameters": [
          {
            "name": "dialog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dialog Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dialogs/{dialog_id}/handoff/force-reset": {
      "post": {
        "tags": [
          "handoff"
        ],
        "summary": "Force Reset Handoff",
        "description": "Force reset handoff state for admin use.\n\n- **dialog_id**: ID of the dialog to reset\n\nOnly available to admin users. Resets dialog to normal state\nregardless of current handoff status. Use carefully!",
        "operationId": "force_reset_handoff_api_dialogs__dialog_id__handoff_force_reset_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dialog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dialog Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/operator/heartbeat": {
      "post": {
        "tags": [
          "operator"
        ],
        "summary": "Operator Heartbeat",
        "description": "Operator heartbeat to maintain presence and update capacity.\n\n- **body**: Status and capacity settings\n- Updates last_heartbeat timestamp and operator status\n- Should be called every 30 seconds by frontend\n- Operators are auto-marked offline if no heartbeat for 90+ seconds",
        "operationId": "operator_heartbeat_api_operator_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorHeartbeatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/operator/queue": {
      "get": {
        "tags": [
          "operator"
        ],
        "summary": "Get Handoff Queue",
        "description": "Get queue of dialogs waiting for operator.\n\nReturns list of dialogs in 'requested' state ordered by request time.\nIncludes last user message, wait time, and priority for each dialog.\nUsed by admin interface to show operator queue.",
        "operationId": "get_handoff_queue_api_operator_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HandoffQueueItem"
                  },
                  "type": "array",
                  "title": "Response Get Handoff Queue Api Operator Queue Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/operator/my-dialogs": {
      "get": {
        "tags": [
          "operator"
        ],
        "summary": "Get My Active Dialogs",
        "description": "Get dialogs currently assigned to this operator.\n\nReturns list of active dialogs assigned to current user.\nUsed by operator interface to show their current workload.",
        "operationId": "get_my_active_dialogs_api_operator_my_dialogs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Get My Active Dialogs Api Operator My Dialogs Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HandoffCancelIn": {
        "properties": {},
        "type": "object",
        "title": "HandoffCancelIn",
        "description": "Schema for cancel request."
      },
      "HandoffQueueItem": {
        "properties": {
          "dialog_id": {
            "type": "integer",
            "title": "Dialog Id",
            "description": "Dialog ID"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At",
            "description": "Request timestamp"
          },
          "reason": {
            "$ref": "#/components/schemas/HandoffReason",
            "description": "Request reason"
          },
          "last_user_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last User Text",
            "description": "Last user message"
          },
          "wait_time_minutes": {
            "type": "integer",
            "title": "Wait Time Minutes",
            "description": "Current wait time in minutes"
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "description": "Priority in queue"
          }
        },
        "type": "object",
        "required": [
          "dialog_id",
          "requested_at",
          "reason",
          "wait_time_minutes",
          "priority"
        ],
        "title": "HandoffQueueItem",
        "description": "Schema for queue item."
      },
      "HandoffReason": {
        "type": "string",
        "enum": [
          "keyword",
          "fallback",
          "retries",
          "manual",
          "auto_trigger"
        ],
        "title": "HandoffReason",
        "description": "Handoff reason enumeration."
      },
      "HandoffReleaseIn": {
        "properties": {},
        "type": "object",
        "title": "HandoffReleaseIn",
        "description": "Schema for release request."
      },
      "HandoffRequestIn": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HandoffReason"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reason for requesting handoff"
          },
          "last_user_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Last User Text",
            "description": "Last message from user"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id",
            "description": "Unique request ID for idempotency"
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "HandoffRequestIn",
        "description": "Schema for handoff request input."
      },
      "HandoffStatus": {
        "type": "string",
        "enum": [
          "none",
          "requested",
          "active",
          "released",
          "cancelled"
        ],
        "title": "HandoffStatus",
        "description": "Handoff status enumeration."
      },
      "HandoffStatusOut": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/HandoffStatus",
            "description": "Current handoff status"
          },
          "assigned_manager": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Manager",
            "description": "Assigned manager info"
          },
          "requested_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested At",
            "description": "When handoff was requested"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At",
            "description": "When handoff was started"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At",
            "description": "When handoff was resolved"
          },
          "reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HandoffReason"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reason for handoff"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Original request ID"
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position",
            "description": "Position in queue"
          },
          "estimated_wait_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Wait Minutes",
            "description": "Estimated wait time"
          },
          "sla_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Deadline",
            "description": "SLA deadline"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HandoffStatusOut",
        "description": "Schema for handoff status output."
      },
      "HandoffTakeoverIn": {
        "properties": {},
        "type": "object",
        "title": "HandoffTakeoverIn",
        "description": "Schema for takeover request."
      },
...
      "OperatorHeartbeatIn": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/OperatorStatus",
            "description": "Operator status"
          },
          "max_active_chats_web": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Active Chats Web",
            "description": "Max web chats"
          },
          "max_active_chats_telegram": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Active Chats Telegram",
            "description": "Max telegram chats"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "OperatorHeartbeatIn",
        "description": "Schema for operator heartbeat input."
      },
      "OperatorStatus": {
        "type": "string",
        "enum": [
          "online",
          "away",
          "offline"
        ],
        "title": "OperatorStatus",
        "description": "Operator presence status enumeration."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/login"
          }
        }
      }
    }
  }
}